backend/server/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
slowapi
pymongo
PyYAML==6.0.1
logger
cachetools
Requests==2.31.0
